+++ topobank/supplib/serializers.py
@@ -1,5 +1,7 @@
 from django.utils.translation import gettext_lazy as _
+from drf_spectacular.utils import extend_schema_field
 from rest_framework import serializers
+from rest_framework.reverse import reverse
 
 
 # From: RomanKhudobei, https://github.com/encode/django-rest-framework/issues/1655
@@ -52,3 +54,84 @@ class StrictFieldMixin:
             raise serializers.ValidationError(errors)
 
         return attrs
+
+
+@extend_schema_field(
+    {
+        "type": "object",
+        "properties": {
+            "id": {"type": "number"},
+            "url": {"type": "string"},
+            "allow": {"enum": ["view", "edit", "full"]},
+        },
+        "required": ["id", "url", "allow"],
+    }
+)
+class PermissionsField(serializers.Field):
+    """
+    A reusable Django REST Framework field that returns a dictionary
+    containing both the object's identifier and a hyperlinked URL.
+
+    The serialized representation takes the form:
+
+        {
+            "id": <object id>,
+            "url": <hyperlinked URL>
+            "allow": <permissions of current user>
+        }
+
+    Parameters
+    ----------
+    view_name : str, optional
+        The name of the DRF view used to generate the hyperlink.
+        Must correspond to a valid URL pattern name in the project.
+        (Default: 'authorization:permission-set-v1-detail')
+    lookup_field : str, optional
+        The name of the model field used for URL lookup.
+        (Default: 'pk')
+    **kwargs
+        Additional keyword arguments passed to the parent `Field` class.
+    """
+
+    def __init__(
+        self,
+        view_name="authorization:permission-set-v1-detail",
+        lookup_field="pk",
+        **kwargs,
+    ):
+        self.view_name = view_name
+        self.lookup_field = lookup_field
+        super().__init__(**kwargs)
+
+    def to_representation(self, obj):
+        """
+        Convert the model instance into a dictionary containing
+        both the object's ID and its hyperlinked URL.
+
+        Parameters
+        ----------
+        obj : Model instance
+            The model instance being serialized.
+
+        Returns
+        -------
+        dict
+            A dictionary with the following structure:
+            {
+                "id": <object id>,
+                "url": <hyperlinked URL>
+                "allow": <permissions of current user>
+            }
+        """
+        request = self.context.get("request", None)
+        lookup_value = getattr(obj, self.lookup_field, None)
+
+        url = None
+        if lookup_value is not None and self.view_name:
+            url = reverse(
+                self.view_name,
+                kwargs={self.lookup_field: lookup_value},
+                request=request,
+            )
+
+        return {"id": lookup_value, "url": url, "allow": obj.get_for_user(request.user)}